refactor base server
